refactor(orchestration): delete progress-reporter facade, retarget importers to sub-modules, baseline-refresh: (refs #3985)
