+++ .agents/scripts/epic-execute-record-wave.js
@@ -42,7 +42,7 @@ import { getRunners } from './lib/config/runners.js';
 import { resolveConfig } from './lib/config-resolver.js';
 import { Logger } from './lib/Logger.js';
 import * as epicRunStateStore from './lib/orchestration/epic-run-state-store.js';
-import { upsertEpicRunProgress } from './lib/orchestration/epic-runner/progress-reporter.js';
+import { upsertEpicRunProgress } from './lib/orchestration/epic-runner/progress-reporter/composition.js';
 import {
   emitStoryDispatchEnd,
   storyStatusToDispatchOutcome,
